refactor : BaseEntity 위치 조정
- global/entity으로 변경
