+++ backend/core/src/main/java/site/timecapsulearchive/core/domain/history/entity/HistoryImage.java
@@ -12,7 +12,7 @@ import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import site.timecapsulearchive.core.domain.BaseEntity;
+import site.timecapsulearchive.core.global.entity.BaseEntity;
 import site.timecapsulearchive.core.domain.capsule.entity.Image;
 
 @Entity
