+++ backend/notification/src/main/java/site/timecapsulearchive/notification/entity/Notification.java
@@ -9,6 +9,7 @@ import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import site.timecapsulearchive.notification.global.entity.BaseEntity;
 
 @Entity
 @Getter
